+++ apps/sim/hooks/queries/tables.ts
@@ -96,33 +96,9 @@ import {
   optimisticallyScheduleNewlyEligibleGroups,
 } from '@/lib/table/deps'
 import { runUploadStrategy } from '@/lib/uploads/client/direct-upload'
+import { type TableQueryScope, tableKeys } from '@/hooks/queries/utils/table-keys'
 
 const logger = createLogger('TableQueries')
-
-type TableQueryScope = 'active' | 'archived' | 'all'
-
-export const tableKeys = {
-  all: ['tables'] as const,
-  lists: () => [...tableKeys.all, 'list'] as const,
-  list: (workspaceId?: string, scope: TableQueryScope = 'active') =>
-    [...tableKeys.lists(), workspaceId ?? '', scope] as const,
-  details: () => [...tableKeys.all, 'detail'] as const,
-  detail: (tableId: string) => [...tableKeys.details(), tableId] as const,
-  exportJobs: (workspaceId?: string) =>
-    [...tableKeys.all, 'export-jobs', workspaceId ?? ''] as const,
-  rowsRoot: (tableId: string) => [...tableKeys.detail(tableId), 'rows'] as const,
-  infiniteRows: (tableId: string, paramsKey: string) =>
-    [...tableKeys.rowsRoot(tableId), 'infinite', paramsKey] as const,
-  rowWrites: (tableId: string) => [...tableKeys.rowsRoot(tableId), 'write'] as const,
-  find: (tableId: string, paramsKey: string) =>
-    [...tableKeys.rowsRoot(tableId), 'find', paramsKey] as const,
-  activeDispatches: (tableId: string) =>
-    [...tableKeys.detail(tableId), 'active-dispatches'] as const,
-  enrichmentDetails: (tableId: string) =>
-    [...tableKeys.detail(tableId), 'enrichment-detail'] as const,
-  enrichmentDetail: (tableId: string, rowId: string, groupId: string) =>
-    [...tableKeys.enrichmentDetails(tableId), rowId, groupId] as const,
-}
 
 type TableRowsParams = Omit<TableRowsQueryInput, 'filter' | 'sort'> &
   TableIdParamsInput & {
